📑 ::내정보 상단 이미지 추가, ⚡️ :내정보 헤더 컴포넌트 스플릿:
